Debug and Utils were moved to new folder Utils
