+++ packages/framework/src/Component/Image/Config/ImageConfigLoader.php
@@ -2,7 +2,7 @@
 
 namespace Shopsys\FrameworkBundle\Component\Image\Config;
 
-use Shopsys\FrameworkBundle\Component\Utils;
+use Shopsys\FrameworkBundle\Component\Utils\Utils;
 use Symfony\Component\Config\Definition\Processor;
 use Symfony\Component\Filesystem\Filesystem;
 use Symfony\Component\Yaml\Parser;
