+++ packages/framework/src/Component/Utils/Utils.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Shopsys\FrameworkBundle\Component;
+namespace Shopsys\FrameworkBundle\Component\Utils;
 
 class Utils
 {
